Reorganize files without functional changes
Mostly disperse the issutil package
(move most of it to subpackages of iss),
split leader selection policies into different files,
and give a few variables more meaningful names.
