+++ pkg/orderers/orderer.go
@@ -9,12 +9,11 @@ package orderers
 import (
 	"fmt"
 
-	"github.com/filecoin-project/mir/pkg/pb/availabilitypb"
-	"github.com/filecoin-project/mir/pkg/pb/messagepb"
-	"github.com/filecoin-project/mir/pkg/util/issutil"
-
+	issconfig "github.com/filecoin-project/mir/pkg/iss/config"
 	"github.com/filecoin-project/mir/pkg/modules"
+	"github.com/filecoin-project/mir/pkg/pb/availabilitypb"
 	"github.com/filecoin-project/mir/pkg/pb/eventpb"
+	"github.com/filecoin-project/mir/pkg/pb/messagepb"
 	"github.com/filecoin-project/mir/pkg/pb/ordererspb"
 
 	"github.com/filecoin-project/mir/pkg/events"
@@ -125,7 +124,7 @@ func NewOrdererModule(
 			proposalTimeout:   0,
 		},
 		messageBuffers: messagebuffer.NewBuffers(
-			issutil.RemoveNodeID(config.Membership, ownID), // Create a message buffer for everyone except for myself.
+			removeNodeID(config.Membership, ownID), // Create a message buffer for everyone except for myself.
 			config.MsgBufCapacity,
 			//       Even better, share the same buffers with ISS.
 			logging.Decorate(logger, "Msgbuf: "),
@@ -137,7 +136,7 @@ func NewOrdererModule(
 	}
 }
 
-func newOrdererConfig(issParams *issutil.ModuleParams, membership []t.NodeID, epochNr t.EpochNr) *PBFTConfig {
+func newOrdererConfig(issParams *issconfig.ModuleParams, membership []t.NodeID, epochNr t.EpochNr) *PBFTConfig {
 
 	// Return a new PBFT configuration with selected values from the ISS configuration.
 	return &PBFTConfig{
@@ -475,6 +474,26 @@ func computeTimeout(timeout t.TimeDuration, view t.PBFTViewNr) t.TimeDuration {
 		view--
 	}
 	return timeout
+}
+
+// removeNodeID removes a node ID from a list of node IDs.
+// Takes a membership list and a Node ID and returns a new list of nodeIDs containing all IDs from the membership list,
+// except for (if present) the specified nID.
+// This is useful for obtaining the list of "other nodes" by removing the own ID from the membership.
+func removeNodeID(membership []t.NodeID, nID t.NodeID) []t.NodeID {
+
+	// Allocate the new node list.
+	others := make([]t.NodeID, 0, len(membership))
+
+	// Add all membership IDs except for the specified one.
+	for _, nodeID := range membership {
+		if nodeID != nID {
+			others = append(others, nodeID)
+		}
+	}
+
+	// Return the new list.
+	return others
 }
 
 // The ImplementsModule method only serves the purpose of indicating that this is a Module and must not be called.
