--- pkg/util/issutil/issutil.go
@@ -1,25 +1,0 @@
-package issutil
-
-import (
-	t "github.com/filecoin-project/mir/pkg/types"
-)
-
-// RemoveNodeID removes a node ID from a list of node IDs.
-// Takes a membership list and a Node ID and returns a new list of nodeIDs containing all IDs from the membership list,
-// except for (if present) the specified nID.
-// This is useful for obtaining the list of "other nodes" by removing the own ID from the membership.
-func RemoveNodeID(membership []t.NodeID, nID t.NodeID) []t.NodeID {
-
-	// Allocate the new node list.
-	others := make([]t.NodeID, 0, len(membership))
-
-	// Add all membership IDs except for the specified one.
-	for _, nodeID := range membership {
-		if nodeID != nID {
-			others = append(others, nodeID)
-		}
-	}
-
-	// Return the new list.
-	return others
-}
